requirements.txt
# Core dependencies
python-jobspy>=1.1.80
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.1.0
numpy==1.26.3
pydantic>=2.3.0
tls-client>=1.0.1
markdownify>=0.13.1
regex>=2024.4.28

# Web framework
Flask>=2.3.0
Flask-CORS>=4.0.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.4.0

# Production dependencies
gunicorn>=21.2.0
supervisor>=4.2.5 